Support search by latitude and longitude coordinates and setting coordinates field by map marker

+++ src/main/java/org/birdup/api/model/entity/Sighting.java
@@ -7,19 +7,16 @@ import com.fasterxml.jackson.annotation.Nulls;
 import java.time.LocalDateTime;
 
 public record Sighting(
-        String speciesCode,
-        String comName,
-        String sciName,
-        String locId,
-        String locName,
-        @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm")
-        LocalDateTime obsDt,
-        @JsonSetter(nulls = Nulls.SKIP)
-        Integer howMany,
-        long lat,
-        long lng,
-        boolean obsValid,
-        boolean obsReviewed,
-        boolean locationPrivate,
-        String subId
-)  {}
+    String speciesCode,
+    String comName,
+    String sciName,
+    String locId,
+    String locName,
+    @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm") LocalDateTime obsDt,
+    @JsonSetter(nulls = Nulls.SKIP) Integer howMany,
+    double lat,
+    double lng,
+    boolean obsValid,
+    boolean obsReviewed,
+    boolean locationPrivate,
+    String subId) {}
